src/bot.py
import os
import openai
import discord
from random import randrange
from src.aclient import client
from discord import app_commands
from src import art, personas, responses

#logger = log.setup_logger(__name__)

def run_discord_bot():
    @client.event
    async def on_ready():
        await client.send_start_prompt()
        await client.tree.sync()

    @client.tree.command(name="chatgpt", description="Chat with chatGPT")
    async def chatGPT(interaction: discord.Interaction, *, message: str):
        if client.is_replying_all == "True":
            await interaction.response.defer(ephemeral=False)
            await interaction.followup.send(
                "")
            return
        if interaction.user == client.user:
            return
        username = str(interaction.user)
        channel = str(interaction.channel)
        await client.send_message(interaction, message)


    @client.tree.command(name="create", description="Generate an image")
    async def create(interaction: discord.Interaction, *, prompt: str):
        if interaction.user == client.user:
            return

        username = str(interaction.user)
        channel = str(interaction.channel)

        await interaction.response.defer(thinking=True, ephemeral=client.isPrivate)
        try:
            path = await art.draw(prompt)

            file = discord.File(path, filename="image.png")
            title = f'> **{prompt}**' + '\n\n'
            embed = discord.Embed(title=title)
            embed.set_image(url="attachment://image.png")

            await interaction.followup.send(file=file, embed=embed)

        except openai.InvalidRequestError:
            await interaction.followup.send(
                "Inappropriate request")

        except Exception as e:
            await interaction.followup.send(
                "Error")


    @client.event
    async def on_message(message):
        if client.is_replying_all == "True":
            if message.author == client.user:
                return
            if client.replying_all_discord_channel_id:
                if message.channel.id == int(client.replying_all_discord_channel_id):
                    username = str(message.author)
                    user_message = str(message.content)
                    channel = str(message.channel)
                   
                    await client.send_message(message, user_message)


    TOKEN = os.getenv("DISCORD_BOT_TOKEN")

    client.run(TOKEN)
